+++ boundlexx/api/v1/views/emoji.py
@@ -2,11 +2,11 @@ from typing import List
 
 from django.db.models import Q
 from django.shortcuts import get_object_or_404
-from rest_framework import filters, viewsets
+from rest_framework import filters
 from rest_fuzzysearch.search import RankedFuzzySearchFilter
 
 from boundlexx.api.common.filters import DedupedFilter
-from boundlexx.api.common.mixins import DescriptiveAutoSchemaMixin
+from boundlexx.api.common.viewsets import BoundlexxViewSet
 from boundlexx.api.utils import get_base_url, get_list_example
 from boundlexx.api.v1.serializers import URLEmojiSerializer
 from boundlexx.boundless.models import Emoji
@@ -18,7 +18,7 @@ EMOJI_EXAMPLE = {
 }
 
 
-class EmojiViewSet(DescriptiveAutoSchemaMixin, viewsets.ReadOnlyModelViewSet):
+class EmojiViewSet(BoundlexxViewSet):
     queryset = Emoji.objects.filter(active=True).prefetch_related("emojialtname_set")
     serializer_class = URLEmojiSerializer
     lookup_field = "name"